+++ packages/tgweb/CHANGELOG.md
@@ -1,5 +1,14 @@
 # tgweb CHANGELOG
 
+## 0.0.9
+
+* 0207b41 Introduce `<tg-template>` elements
+* ffbd295 Introduce `<tg-link>` elements
+* 5c9689d Introduce `<tg-component>` and `<tg-insert>` elements
+* e8ce67a Introduce `<tg-article>`, `<tg-articles>` and `<tg-links>` elements
+* b868d37 Introduce `<tg-if-complete>` elements
+* ae8c331 Improve the manipulation of `<tg-link>` elements
+
 ## 0.0.7
 
 * 4bafb1b Make `tg-article`, `tg-articles` and `tg-links` accept slots